Intoroduce LogInKit class.
